==> src/1 - treating_df.ipynb <==
import pandas as pd
import os
import numpy as np
import path_relative as po

print(os.getcwd())


training_columns = ['RSRP', 'RSRQ', 'SNR', 'CQI', 'RSSI']

### --------------------------  Os datasets novos  ----------------------------###
netflix_static_dataset = pd.DataFrame()
netflix_driving_dataset = pd.DataFrame()
amazon_static_dataset = pd.DataFrame()
amazon_driving_dataset = pd.DataFrame()

### -------------------------- Paths of datasets ------------------------------ #####
output_folder_netflix_static = r'C:\Users\IvoAg\OneDrive\Área de Trabalho\Repositorios\AI_applied_to_5g\5G - Datasets treated\Netflix\Static'
output_folder_netflix_driving = r'C:\Users\IvoAg\OneDrive\Área de Trabalho\Repositorios\AI_applied_to_5g\5G - Datasets treated\Netflix\Driving'
output_folder_amazon_static = r'C:\Users\IvoAg\OneDrive\Área de Trabalho\Repositorios\AI_applied_to_5g\5G - Datasets treated\Amazon\Static'
output_folder_amazon_driving = r'C:\Users\IvoAg\OneDrive\Área de Trabalho\Repositorios\AI_applied_to_5g\5G - Datasets treated\Amazon\Driving'


#### ----------------------------- Filenames of the datasets --------------------------------- ###
output_filename_netflix_static = 'netflix_static.csv'
output_filename_netflix_driving = 'netflix_driving.csv'
output_filename_amazon_static = 'amazon_static.csv'
output_filename_amazon_driving = 'amazon_driving.csv'



### Concatenate Netflix dataset staitc ###


for p in po.path_Netflix_Static_Season3_StrangerThings:
    path =os.path.join('..',p)
    data = pd.read_csv(path)
    data = data.replace("-", np.nan)
    median_value = data['RSSI'].median()
    data['RSSI'].fillna(median_value, inplace=True)
    netflix_static_dataset =  pd.concat([netflix_static_dataset, data], ignore_index=True) 

for p in po.path_Netflix_Static_animated_RickandMorty :
    path =os.path.join('..',p)
    data = pd.read_csv(path)
    data = data.replace("-", np.nan)
    median_value = data['RSSI'].median()
    data['RSSI'].fillna(median_value, inplace=True)
    netflix_static_dataset =  pd.concat([netflix_static_dataset, data], ignore_index=True)


netflix_static_dataset.to_csv(output_folder_netflix_static +  '\\'+ output_filename_netflix_static, index=False) 




    

### Concatenate Netflix driving dataset ### 



for p in po.path_Netflix_Driving_animated_RickandMorty:
    path =os.path.join('..',p)
    data = pd.read_csv(path)
    data = data.replace("-", np.nan)
    median_value = data['RSSI'].median()
    data['RSSI'].fillna(median_value, inplace=True)
    netflix_driving_dataset =  pd.concat([netflix_driving_dataset, data], ignore_index=True) 

for p in po.path_Netflix_Driving_Season3_StrangerThings :
    path =os.path.join('..',p)
    data = pd.read_csv(path)
    data = data.replace("-", np.nan)
    median_value = data['RSSI'].median()
    data['RSSI'].fillna(median_value, inplace=True)
    netflix_driving_dataset =  pd.concat([netflix_driving_dataset, data], ignore_index=True)


netflix_driving_dataset.to_csv(output_folder_netflix_driving +  '\\'+ output_filename_netflix_driving, index=False) 


### Concatenate Amazon static dataset ### 

for p in po.path_AmazonPrime_Static_animatedNinjago:
    path =os.path.join('..',p)
    data = pd.read_csv(path)
    data = data.replace("-", np.nan)
    median_value = data['RSSI'].median()
    data['RSSI'].fillna(median_value, inplace=True)
    amazon_static_dataset =  pd.concat([amazon_static_dataset, data], ignore_index=True) 

for p in po.path_AmazonPrime_Static_Season3TheExpanse :
    path =os.path.join('..',p)
    data = pd.read_csv(path)
    data = data.replace("-", np.nan)
    median_value = data['RSSI'].median()
    data['RSSI'].fillna(median_value, inplace=True)
    amazon_static_dataset =  pd.concat([amazon_static_dataset, data], ignore_index=True)


amazon_static_dataset.to_csv(output_folder_amazon_static +  '\\'+ output_filename_amazon_static, index=False) 

### Concatenate Amazon driving dataset ### 

for p in po.path_AmazonPrime_Driving_AnimatedAdventureTime:
    path =os.path.join('..',p)
    data = pd.read_csv(path)
    data = data.replace("-", np.nan)
    median_value = data['RSSI'].median()
    data['RSSI'].fillna(median_value, inplace=True)
    amazon_driving_dataset =  pd.concat([amazon_driving_dataset, data], ignore_index=True) 

for p in po.path_AmazonPrime_Driving_Season3TheExpanse:
    path =os.path.join('..',p)
    data = pd.read_csv(path)
    data = data.replace("-", np.nan)
    median_value = data['RSSI'].median()
    data['RSSI'].fillna(median_value, inplace=True)
    amazon_driving_dataset =  pd.concat([amazon_driving_dataset, data], ignore_index=True)


amazon_driving_dataset.to_csv(output_folder_amazon_driving +  '\\'+ output_filename_amazon_driving, index=False)
