==> src/2 - generate_training_data.ipynb <==
import pandas as pd
import numpy as np
from sklearn.model_selection import train_test_split


#### Import the data set and check the NaN values ####

dataset = pd.read_csv(r'../5G - Datasets treated/Amazon/Driving/amazon_driving.csv')
columns_to_fill = ['Longitude', 'Latitude','Speed','RSRP', 'RSRQ', 'SNR', 'CQI', 'RSSI']
medians = dataset[columns_to_fill].median()
dataset[columns_to_fill] = dataset[columns_to_fill].fillna(medians)
nan_counts = dataset.isna().sum()

print("Number of NaN values in each column:")
print(nan_counts)


#### Fill the NaN colluns with median  (meidana) if neccessary ####

dataset[columns_to_fill] = dataset[columns_to_fill].fillna(medians)
print("Number of NaN values in each column:")
print(nan_counts)

print(dataset.columns)

#### Saving the training and test data ####

X_DL = dataset[['Longitude', 'Latitude','Speed','RSRP', 'RSRQ', 'SNR', 'CQI', 'RSSI']]
y_DL = dataset['DL_bitrate']

X_UL = dataset[['Longitude', 'Latitude','Speed','RSRP', 'RSRQ', 'SNR', 'CQI', 'RSSI']]
y_UL = dataset['UL_bitrate']

X_train_DL, X_test_DL, y_train_DL, y_test_DL = train_test_split(X_DL, y_DL, test_size=0.3, random_state=42)
X_train_UL, X_test_UL, y_train_UL, y_test_UL = train_test_split(X_UL, y_UL, test_size=0.3, random_state=42)

output_folder_train_and_test = r'../Training_And_Test/Amazon/'
X_train_DL.to_csv(output_folder_train_and_test + '//' + 'X_train_DL_driving.csv')
y_train_DL.to_csv(output_folder_train_and_test + '//' + 'y_train_DL_driving.csv')
X_test_DL.to_csv(output_folder_train_and_test + '//' + 'X_test_DL_driving.csv')
y_test_DL.to_csv(output_folder_train_and_test + '//' + 'y_test_DL_driving.csv')

X_train_UL.to_csv(output_folder_train_and_test + '//' + 'X_train_UL_driving.csv')
y_train_UL.to_csv(output_folder_train_and_test + '//' + 'y_train_UL_driving.csv')
X_test_UL.to_csv(output_folder_train_and_test + '//' + 'X_test_UL_driving.csv')
y_test_UL.to_csv(output_folder_train_and_test + '//' + 'y_test_UL_driving.csv')
