==> src/3 - hyperparameters_choice.ipynb <==
import pandas as pd
import numpy as np
from sklearn.model_selection import train_test_split
from sklearn import metrics
from sklearn.model_selection import cross_val_score
from sklearn.preprocessing import StandardScaler
from sklearn.pipeline import Pipeline
import joblib

X_test = pd.read_csv(r'../Training_And_Test/Amazon/X_test_DL_static.csv')
y_test = pd.read_csv(r'../Training_And_Test/Amazon/y_test_DL_static.csv')
X_train = pd.read_csv(r'../Training_And_Test/Amazon/X_train_DL_static.csv')
y_train = pd.read_csv(r'../Training_And_Test/Amazon/y_train_DL_static.csv')

### Random Forest Regressor  ###


from sklearn.ensemble import RandomForestRegressor
from sklearn.model_selection import RandomizedSearchCV
from scipy.stats import randint

#### Randomized Search ####

rf_reg = RandomForestRegressor()


param_dist = {
    'n_estimators': randint(100, 2000)  
}

random_search = RandomizedSearchCV(
    estimator=rf_reg,
    param_distributions=param_dist,
    n_iter=12,
    cv=2,
    scoring='neg_mean_squared_error',
    verbose=2,
    n_jobs= 5
)

pipeline = Pipeline([
    ('scaler', StandardScaler()),  
    ('random_search', random_search),
])

pipeline.fit(X_train, y_train)



print("Best Parameters:", pipeline.named_steps['random_search'].best_params_)


rf_reg_random_search = pipeline
joblib.dump(rf_reg_random_search, r'../Models/Amazon/Driving/Upload/random_forest_resgressor_random_search.pkl')

#### Grid Search ####

from sklearn.model_selection import GridSearchCV
from sklearn.ensemble import RandomForestRegressor


rf_reg = RandomForestRegressor()

param_grid = {
    'n_estimators': [100, 150, 200, 300, 400 ,500, 700, 1000,1200, 1500, 1800, 2000]
}

grid_search = GridSearchCV(
    estimator=rf_reg, 
    param_grid=param_grid, 
    cv=2, 
    scoring='neg_mean_squared_error',
    n_jobs= 5,
    verbose=2,
)


pipeline = Pipeline([
    ('scaler', StandardScaler()),  
    ('grid_search', grid_search),
])


pipeline.fit(X_train, y_train)

print("Best Parameters:", pipeline.named_steps['grid_search'].best_params_)


rf_reg_grid_search = pipeline
joblib.dump(rf_reg_grid_search , r'../Models/Amazon/Driving/Upload/random_forest_resgressor_grid_search.pkl')

### Suporte Vector Machine ###

from sklearn.svm import SVR
import matplotlib.pyplot as plt
import seaborn as sns
from sklearn.model_selection import cross_val_score
from sklearn.model_selection import GridSearchCV
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns

#### Grid Search ####

from sklearn.model_selection import GridSearchCV
from sklearn.svm import SVR

svm_reg = SVR(C = 1000,gamma = 0.0001 )

param_grid = {
    'kernel': [ 'rbf', 'poly']   
}


grid_search = GridSearchCV(
    estimator= svm_reg, 
    param_grid=param_grid, 
    cv=2, 
    scoring='neg_mean_squared_error',
    verbose=2,
)


pipeline = Pipeline([
    ('scaler', StandardScaler()),  
    ('grid_search', grid_search),
])


pipeline.fit(X_train, y_train['UL_bitrate'])

best_params = pipeline.named_steps['grid_search'].best_params_
print("Best Parameters:", best_params)

svr_grid_search = pipeline
joblib.dump(svr_grid_search , r'../Models/Amazon/Driving/Upload/suport_vector.pkl')

### MultiLayer Perceptron ###


import numpy as np
from sklearn.model_selection import train_test_split
from sklearn.preprocessing import StandardScaler
from sklearn.neural_network import MLPRegressor
from sklearn.pipeline import Pipeline
from sklearn.model_selection import GridSearchCV, RandomizedSearchCV
from sklearn.metrics import mean_squared_error

# Create a pipeline with StandardScaler and MLPRegressor
pipeline = Pipeline([
    ('scaler', StandardScaler()),  # Standardize features
    ('mlp', MLPRegressor(random_state=42))  # MLP Regressor
])

# Define the hyperparameter grid for Grid Search
param_grid = {
    'mlp__hidden_layer_sizes': [(50, 25), (100, 50), (100, 100)],
    'mlp__activation': ['relu', 'tanh'],
    'mlp__alpha': [0.0001, 0.001, 0.01]
}

# Perform Grid Search with cross-validation
grid_search = GridSearchCV(
    estimator=pipeline, 
    param_grid=param_grid, 
    cv=5, 
    scoring='neg_mean_squared_error', 
    n_jobs= 5,
    verbose = 2
    )
grid_search.fit(X_train, y_train)

# Get the best parameters and estimator from Grid Search
best_params = grid_search.best_params_
best_estimator = grid_search.best_estimator_

# Train the model with the best parameters on the full training set
best_estimator.fit(X_train, y_train)

# Make predictions on the test set
y_pred = best_estimator.predict(X_test)

# Evaluate the model
mse = mean_squared_error(y_test, y_pred)
print(f"Mean Squared Error (MSE): {mse}")

# Alternatively, perform Randomized Search with cross-validation
random_search = RandomizedSearchCV(
    estimator=pipeline, 
    param_distributions=param_grid, 
    n_iter=10, 
    cv=5, 
    scoring='neg_mean_squared_error', 
    n_jobs=4,
    verbose = 2
    )
random_search.fit(X_train, y_train)

# Get the best parameters and estimator from Randomized Search
best_params_random = random_search.best_params_
best_estimator_random = random_search.best_estimator_

# Train the model with the best parameters from Randomized Search on the full training set
best_estimator_random.fit(X_train, y_train)

# Make predictions on the test set
y_pred_random = best_estimator_random.predict(X_test)

# Evaluate the model
mse_random = mean_squared_error(y_test, y_pred_random)
print(f"Mean Squared Error (MSE) with Randomized Search: {mse_random}")

# Compare os MSEs dos modelos
if mse < mse_random:
    best_model = best_estimator
    best_params = best_params
    print("O modelo do Grid Search é melhor.")
    print(f"Mean Squared Error (MSE) do Grid Search: {mse}")
elif mse_random < mse:
    best_model = best_estimator_random
    best_params = best_params_random
    print("O modelo do Randomized Search é melhor.")
    print(f"Mean Squared Error (MSE) do Randomized Search: {mse_random}")
else:
    best_model = best_estimator  # Use o modelo do Grid Search por padrão
    best_params = best_params
    print("Os modelos do Grid Search e Randomized Search têm desempenhos semelhantes.")
    print(f"Mean Squared Error (MSE) do Grid Search: {mse}")
    print(f"Mean Squared Error (MSE) do Randomized Search: {mse_random}")

# Imprima os melhores parâmetros do melhor modelo
print("Melhores Parâmetros do Melhor Modelo:")
print(best_params)

# Agora você pode usar 'best_model' para fazer previsões ou qualquer outra operação necessária.
