==> src/4 - models.ipynb <==
### Imports Default ###

import pandas as pd
import numpy as np
from sklearn.model_selection import train_test_split
from sklearn import metrics
from sklearn.model_selection import cross_val_score
from sklearn.preprocessing import StandardScaler
from sklearn.pipeline import Pipeline
import joblib


X_test = pd.read_csv(r'../Training_And_Test/Netflix/X_test_DL_static.csv')
y_test = pd.read_csv(r'../Training_And_Test/Netflix/y_test_DL_static.csv')
X_train = pd.read_csv(r'../Training_And_Test/Netflix/X_train_DL_static.csv')
y_train = pd.read_csv(r'../Training_And_Test/Netflix/y_train_DL_static.csv')

### Linear Regression ###

from sklearn.linear_model import LinearRegression as LR

# Create a pipeline
pipeline = Pipeline([
    ('scaler', StandardScaler()),  # Data scaling step
    ('model', LR())  # Linear Regression model step
])


pipeline.fit(X_train, y_train)

linear_model = pipeline
joblib.dump(linear_model, r'../Models/Netflix/Static/Dowload/linear.pkl')

### Random Forest Regressor ###

from sklearn.ensemble import RandomForestRegressor

pipeline = Pipeline([
    ('scaler', StandardScaler()), 
    ('rf', RandomForestRegressor(n_estimators=1561, random_state=42))  # Random Forest Regressor
])

# Fit the pipeline to the training data
pipeline.fit(X_train, y_train)


random_forest_regressor_randomized_search = pipeline

from sklearn.ensemble import RandomForestRegressor

pipeline = Pipeline([
    ('scaler', StandardScaler()), 
    ('rf', RandomForestRegressor(n_estimators=200, random_state=42))  # Random Forest Regressor
])

# Fit the pipeline to the training data
pipeline.fit(X_train, y_train)

random_forest_regressor_grid_search = pipeline

from sklearn.metrics import mean_squared_error

y_pred_grid = random_forest_regressor_grid_search.predict(X_test)
y_pred_random = random_forest_regressor_randomized_search.predict(X_test)

# Calcular o MSE para ambos os modelos
mse1 = mean_squared_error(y_test, y_pred_grid)
mse2 = mean_squared_error(y_test, y_pred_random)

# Comparar os MSEs
if mse1 < mse2:
    print("O primeiro modelo tem um MSE menor e é considerado melhor.")
elif mse2 < mse1:
    print("O segundo modelo tem um MSE menor e é considerado melhor.")
else:
    print("Ambos os modelos têm resultados semelhantes em termos de MSE.")

### Escolhodo o melhor eu salvo aqui
joblib.dump(random_forest_regressor_randomized_search, r'../Models/Netflix/Static/Dowload/random_forest_regressor.pkl')

Support Vector Regression

from sklearn.svm import SVR

pipeline = Pipeline([
    ('scaler', StandardScaler()),  # Standardize features
    ('svr', SVR( C = 1000, gamma =  0.0001, kernel='poly'))  # Support Vector Regressor with linear kernel
])

# Fit the pipeline to the training data
pipeline.fit(X_train, y_train['DL_bitrate'])

svr = pipeline
joblib.dump(svr, r'../Models/Netflix/Static/Dowload/support_vector_regression.pkl')

### Multilayer Perceptron ###
