==> src/5 -plots.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
import hvplot.pandas
from sklearn import metrics
from sklearn.model_selection import cross_val_score
import joblib

### Test and Models in the we will use ###

### ---------------------------------- Netflix --------------------------- ###
y_test_netflix_static_DL = pd.read_csv(r'../Training_And_Test/Netflix/y_test_DL_static.csv')
y_test_netflix_static_UL = pd.read_csv(r'../Training_And_Test/Netflix/y_test_UL_static.csv')
y_test_netflix_driving_DL = pd.read_csv(r'../Training_And_Test/Netflix/y_test_DL_driving.csv')
y_test_netflix_driving_UL = pd.read_csv(r'../Training_And_Test/Netflix/y_test_UL_driving.csv')

### ----------------------------------- Amazon --------------------------- ####

y_test_amazon_static_DL = pd.read_csv(r'../Training_And_Test/Amazon/y_test_DL_static.csv')
y_test_amazon_static_UL = pd.read_csv(r'../Training_And_Test/Amazon/y_test_UL_static.csv')
y_test_amazon_driving_DL = pd.read_csv(r'../Training_And_Test/Amazon/y_test_DL_driving.csv')
y_test_amazon_driving_UL = pd.read_csv(r'../Training_And_Test/Amazon/y_test_UL_driving.csv')



#### Linear Model ####


linear_model_amazon_driving_DL = joblib.load(r'../Models/Amazon/Driving/Dowload/linear.pkl')
linear_model_amazon_driving_UL = joblib.load(r'../Models/Amazon/Driving/Upload/linear.pkl')
linear_model_amazon_static_DL = joblib.load(r'../Models/Amazon/Static/Dowload/linear.pkl')
linear_model_amazon_static_UL = joblib.load(r'../Models/Amazon/Static/Upload/linear.pkl')

linear_model_netflix_driving_DL = joblib.load(r'../Models/Netflix/Driving/Dowload/linear.pkl')
linear_model_netflix_driving_UL = joblib.load(r'../Models/Netflix/Driving/Upload/linear.pkl')
linear_model_netflix_static_DL = joblib.load(r'../Models/Netflix/Static/Dowload/linear.pkl')
linear_model_netflix_static_UL = joblib.load(r'../Models/Netflix/Static/Upload/linear.pkl')

#### Random Forest Regressor Grid Search ####

rf_regressor_model_grid_search_amazon_driving_DL = joblib.load(r'../Models/Amazon/Driving/Dowload/random_forest_resgressor_grid_search.pkl')
rf_regressor_model_grid_search_amazon_driving_UL = joblib.load(r'../Models/Amazon/Driving/Upload/random_forest_resgressor_grid_search.pkl')
rf_regressor_model_grid_search_amazon_static_DL = joblib.load(r'../Models/Amazon/Static/Dowload/random_forest_resgressor_grid_search.pkl')
rf_regressor_model_grid_search_amazon_static_UL = joblib.load(r'../Models/Amazon/Static/Upload/random_forest_resgressor_grid_search.pkl')

rf_regressor_model_grid_search_netflix_driving_DL = joblib.load(r'../Models/Netflix/Driving/Dowload/random_forest_resgressor_grid_search.pkl')
rf_regressor_model_grid_search_netflix_driving_UL = joblib.load(r'../Models/Netflix/Driving/Upload/random_forest_resgressor_grid_search.pkl')
rf_regressor_model_grid_search_netflix_static_DL = joblib.load(r'../Models/Netflix/Static/Dowload/random_forest_resgressor_grid_search.pkl')
rf_regressor_model_grid_search_netflix_static_UL = joblib.load(r'../Models/Netflix/Static/Upload/random_forest_resgressor_grid_search.pkl')
